fix get_user_choice for other sides
as the comment in replay.cpp says we want "send data to others" so
calling "pull_remote_user_input" is not really correct. Since
plamp_controller::handle_generaic_command("ai_user_input")
also invokes send_data we just don't call it with "sync_network" in this
case. This solution is still not optmial but it has the smallest chance
to break things since epecialy because this should go to 1.12 too.

Conflicts:
	src/ai/manager.cpp