disable all tests on gcc builds, to prevent timeouts
It seems that we cannot avoid the segfault without adding gcc-4.8
download and install to the test sequence, but this seems to push
us over the time limit consistently. In this commit, go back to
using gcc-4.8 but disable everything but the wesnoth target itself
for the gcc build to save time.