delete the whiteboard
This commit deletes the whiteboard source code, then fixes up
the various references and project files. This is simpler than
deleting references piecemeal, because it turns out that the
whiteboard code was used sublty in many different places and
could be activated or triggered by hotkeys in several different
ways.