remove mp_game_settings::share_view/maps
This is a leftover from 1.0 where this was adjustable in a dialog. This
is not the case anymore, the current behavior results in
[side]share_view= always being true in mp which just means we have one
less option for mp scenario designers.
Even worse, the previous behavior results in share_view always being
false for scenarios in mp campaign that are not the first scenario,
regardless of the settings in the scenario.