campaignd: Use find() to locate a request handler
Makes it so for each request received we use
request_handlers_table::find() instead of a plain linear search to find
the handler, using the tag name of the first WML child node we can find,
instead of trying to locate one with a known handler id.

In theory this allows taking advantage of the request_handlers_table
type's (currently a std::map) own lookup mechanism.