campaignd: Reject bad requests with an [error] block
This can be used to signal campaignd clients in the future about
features that aren't yet implemented.

As with the previous change to using find() to locate a request handler,
this only applies to the first WML child of the request's root node.
Other nodes continue to be ignored.

(Backported efcf8f7b6aa4e25c0ed5cbd80597e2d894658d95 from master.)