fixup 0f5876fe38067c9783f41ab769e485dd8705e30e
There was a small mistake here: The unit constructor
which takes a unit type first and 3 other parameters,
has the 4th one being an optional gender, and the
third a bool. fake_unit had a constructor mapping to
this, which sets the third to false. I over looked
this when factoring out fake_unit, but it only
generates a warning to cast enum to bool on MSVC
compilers.