Fix the linking error found in the last commit.
The error was caused by forgetting to add the new file to the
CMakeList.txt file.  It was then found that the new file did not
compile because of non matching variable names, those were changed
so they match.  Lastly a comment asking for reveiw/help finding the
issue was removed from unit_types.cpp as it is no longer an issue.