Demote "skipping textdomain" error to warning level
The "no location found for '<path>', skipping textdomain" error message
is usually inoffensive and merely means that there is a textdomain
declaration pointing to a directory that doesn't exist, so it shouldn't
really be an error. Since we don't have anything midway between warning
and info messages, this should remain in place as a warning for WML
coders typoing textdomain declarations or forgetting to include their
add-on's translations with it.