mp create engine uses Team + # syntax when use_map_settings=false
The old behavior was that when use_map_settings=false, every side
would have team=# for some number. This causes problems because
config might secretly decide that that is an int, and cause string
comparisons to fail.

Whether or not that is a bug is a separate issue, it seems clearly
desirable for sides to always have a team name of the form
"Team #", to side-step the potential problems.

This fixes up problems which appeared after
4d73031680b8889ec70e42607016269f7c421da5
852b6981329a96055249c513477be0d43b2fcaa0