Consider <exe loc or cwd>/../ as a candidate for data dir autodetection
Makes it so <exe location>/../ (non-Windows) or <current working
dir>/../ (Windows) are also considered possible data directories if they
contain a data/_main.cfg file.

This is mostly intended to help people running cmake builds so they
don't have to type out .. in the Wesnoth command line every time or use
a hard link sitting on the data dir.

Notice that scons people already benefited from the simpler version of
this autodetection, which only considered <exe loc or cwd>/./ as a
candidate. This has been the case since version 1.5.4 (more
specifically, commit 43431bc3a79d8be4fef2c102167e5982a74aa71a). Because
of this, most of the code for this feature was already in place and I
only rearranged it to add some helpful annotations.