add lazy parsing of comma separated lists to unit_filter
The old code seemed to desperately want to avoid calling split
on large lists of strings repeatedly, however it did this again
and again for every unit. The new implementation remembers the old
splits for the next time the filter is applied, but makes sure
only to split them if the old code would have split them.