addon/mg: Include Update All on all views and enable it when applicable
This makes it so the Update All button appears on all views, disabled by
default unless there is at least one ADDON_INSTALLED_UPGRADABLE add-on.

The rationale for this is that with the Update Add-ons shortcut in the
Connect dialog gone, the possibility of upgrading installed add-ons is
not as visible as it was in versions 1.5.3 through 1.10.x. Making the
button visible at all times should hint at the existence of this
functionality.