addon/mg: Choose "update" vs. "install" terms when using Update All too
Previously the "update" terminology was only used when the Upgradable
view was the active view. Now that the Update All action may be
available on other views, we need to take care to use the correct terms
for it too.

This commit incurs in no changes to existing translatable strings.