bugfix for #22285
I haven't decided what the ideal solution is for this yet, might
want to rearrange the play_controller init sequence. But for now,
if the game display is null and we need to make an animation, we
use the game state as a filter context rather than the display,
which for all intents and purposes in current code with be the
same. This looks to fix the bug.