Simplify --resolution error handling
Reuse the boost::program_options::error class (which we already catch in
main()) instead of throwing a std::string, and simplify the throwing
logic accordingly.

There is no need to provide incredibly verbose error messages since the
user is expected to be able to read --help or the man page to get here
in the first place.