revert PR 98
This commit removes the "carryover_wml" feature which was added
just before 1.12 was branched off.

I added this feature at a time when I didn't have as good an
understanding as I do now of the networked mp code, I now believe
that it was poorly concieved and should not be implemented this
way if at all. There are better ways to achieve what was meant to
be achieved here, we are better off to remove it as it currently
exists.