revert PR 98
This is a parallel commit to PR 243, on master instead of 1.12
branch.

The commit is significantly different because this feature got
swept up in a different refactor of playcampaign / carryover. I
didn't test the feature extensively on this branch, this might not
be necessary at all. I'm simply making the sister commit anyways
so that we have something concrete to discuss when deciding
whether to merge either or both of these pull requests.

I do think the feature is of questionable value and adds
complexity to an already confusing part of the engine, and should
therefore be considered for removal.