fix bug 22086, by storing map data correctly in savegame summaries
The bug was that minimap previews for save games were not appearing,
or that the wrong ones were appearing.

It seems that because of more relics from this commit / related,

https://github.com/wesnoth/wesnoth/commit/c17d84b12

the savegame summaries generated by "extract savegame summary" were
bugged and didn't store the map data correctly. This bug wasn't
obvious because the save game load dialog tries to scavenge what
ever the current game config is for map data if it can't find this,
but all of the saveindex data was being generated incorrectly.

After this commit all the save files seem to work correctly,
except for start-of-scenario saves for reasons that aren't clear
to me, and which I think is a separate issue. (In that case,
something is going wrong with the game_config lookup process most
likely.)

This only appears to fix the bug for normal saves, autosaves, and
replay saves.