check lobby_sounds preference for the GUI 1 lobby also
This fixes a bug reported on forums here:
http://forums.wesnoth.org/viewtopic.php?f=4&t=40660#p572131

The lobby sounds pref now affects the GUI 1 and 2 mp lobbies,
but it only affects the lobby screen, not mp::wait or mp::connect
etc.

This check is implemented by trying to dynamic cast 'this' mp::ui
object to a mp::lobby, at the time when we would play the arrive
or leave sounds.

The lobby sounds pref does *not* block the sounds due to server
messages, messages from a friend, or whispers.