give a warning when notifications enabled but not available
The "disable notifications" option in advanced preferences allows
the user to toggle desktop notifications on and off. However, if
the game was not compiled with support for these, the option won't
do anything.

In this revision, if disable_notifications is set to false, but
notifications are not available, a dialog is launched which
explains that this executable was not compiled with support for
notifications.

The preference *is* toggled however -- this is because all versions
of wesnoth use the same preference file (at least under linux), so
we would rather to allow the user to change the file at will --
a different wesnoth binary which they have may indeed support
notifications. So we only give a warning but allow them to toggle.