add "Defense Capped" col. to terrain table in help page for units
If a unit has a "negative" modifier on one of its terrain defenses,
for the basic terrain defense types that have been discovered and
will be displayed on the current help page, a column will be
generated "Defense Capped" which indicates which terrain types
have the negative modifier. In each row the column contains either
'-' or 'yes'.