Add defense caps for mushroom terrain type, tweak some fluff
This commit adds negative defense modifiers, (thus establishing
a defense cap for mushroom mixed terrain types) for the following
units:

- Mountainfoot units (includes dwarves, and incidentally giant
spider and ant)
- Gryphon Rider and Gryphon Master
- Mounted type (includes loyalist cavalry)
- Khalifate mounted and armored mounted

Essentially we are adding it for most units which have very low
defense ratings on mushrooms, so that their mixed terrain defense
ratings will not improve.

The fluff description of mushroom groves is changed in the following
way:

- Mushroom groves are no longer described as "vast underground
forests of mushrooms" because this doesn't make much sense, in the
game they also appear on the surface and not underground. So the
fluff allows that they may appear on the surface as well.
- Dwarves are explained to have particularly low defense on
mushrooms, because mushrooms have an intoxicating effect on dwarves,
and the spore clouds of mature giant mushrooms render them sluggish.
This further explains why gryphon riders are slow over the mushrooms
but gryphons without a dwarf rider are not.