add description of "Best/worst defense/movement" to terrain help
When viewing a terrain description (ctrl-t), or looking at the
description of a mixed terrain in the help, wesnoth will parse
the defense and movement alias strings, and format it in the
translated language, color coding the words "Best of " in green
and "Worst of " in red. These descriptions become parenthesized
with subexpressions if there is a complicated terrain with many
types and multiple alternations.