add remarks about the illuminated underground Times of Day to help
I'm not sure if this stuff actually belongs on this page, because
this page is very fundamental, and this is added detail that may
distract of confuse a beginner. But I think these remarks belong
somewhere in the help, just so that the pictures are explained,
even if this page isn't the right place.