experimental implementation of sp advanced settings dialog
Era and mod selection is functional, as well as options to enable mp_configure and mp_create dialogs.
There is still an issue of realoading config. If you load multiplayer, then exit, then load campaign, there is no config reload, so the multiplayer eras/mods stay in config and are selectable. However, since state.classification().campaign_type is not set to multiplayer, when config is reloaded for a specific campaign, multiplayer won't be defined and therefore the selected era may no longer be available. So, we need some way to handle this case better. Also, it may be desirable to add a button to advanced settings which would change the campaign_type to multiplayer and reload (if we actually did want to load some mp era). Another potential option for this dialog could be IO type (to make the game networked, etc).