fix copy assignment operator for unit class to work with refcount
This issue was spotted by gfgtdf. It doesn't appear that we
actually use this function -- prior to this commit, I tested a
version which had "assert(false)" at the start of this function,
and it was never triggered in normal play. But it's possible that
some lua command or obscure WML tag does use it.

It might also be a good idea to make the unit class noncopyable.