geography.cfg: correct wmllint spelling for Castelfranc and Rumyr
These towns were misspelled in their encyclopedia entries, compared to
the Wesnoth map. All of the instances in my UMC collection follow the map
spellings, so I am not leaving the incorrect spelling as an alternative.

Also, there were spelling rules for both "Memog" and Melmog on separate
lines. No reason to keep the mistake around.