add multiplayer proxy object for lua scripts
This commit adds the "multiplayer" proxy object to the lua
environment. It also adds an underlying "lua_mp_interface" object,
which acts as an intermediary between lua and C++.

- The mp interface object stores a queue of lua requests, which
are generated by the lua script and stored rather than executed
immediately.
- The mp interface object tracks the pointers to the active mp::ui
objects, avoiding the need to do this in lua and simplifying the
lua scripts.

The reason for the interface object is, we don't want the lua api
here to be tightly bound the implementation of these mp::ui's,
because the mp::ui's are slated to change anyways and become gui2.
Having the interface means that, a programmer can make big changes
to the mp::ui's and as long as they can still find ways to support
the script requests, they won't break any of the scripts.

The reason to use a request queue rather than to immediately
execute C++ callbacks when lua makes a request is essentially just
for ease of debugging later. With this system, when a script is
run we can dump the queued lua requests when we need to debug it,
and in general the stack traces from gdb won't have lua functions
like lua_pcall etc. from the lua interpretter, so it will be a
bit friendlier.