make "unit has empty type" error not fatal when starting a game
This can happen easily in a campaign if there are saveid mismatches
and such, and generally the only error message will be "game error:
created a unit with empty type". This commit makes the error not
fatal, and simply skips the unit. Also we give a better error
message, on the warning channel, which tells the id of the unit
and the details of the side which owned it.