add try_fetch_addon_with_checks fcn, use when gui installs addons
Extending the 'try_fetch_addon' fcn, this fcn also calls the dep
check and the check for dev overwrite functions beforehand. Thus it
does a fully 'smart' download of an add-on and tidies up many of the
details.

We plan to call this function also as part of the code path when we
do addon auto-downloads from the mp lobby.