add ad_hoc_addon_fetch_session fcn, installs an add-on given its id
This function is public and placed in the header, so that it may be
called by the mp lobby code.

Essentially, it just takes the client creation and connection
routines from addon_manager_ui, calls the try_fetch_addon_with_checks
function, and wraps this in the exception handlers from
addon_manager_ui.