in game cfg mgr, annotate era, mod tags with the addon_id
This is done at the time of loading the config of all addons. The
addon_id can be deduced without much difficulty from the path, we
do this calculation, then scan over the entire cfg that resulted
from that addon and add an addon_id attribute to appropriate top-
level tags.

The purpose of this is so that this info will be included in a
level config after it is constructed by the host, and can be made
available by connect engine to clients that wish to join.

It might be better to do this even earlier in the pipeline, i.e.
so that it happens before the contents are cached. At the moment
I don't see a compelling reason to do that though, and this way
was fairly simple.