in fcn expand_mp_events, also collect addon_ids, for mp_settings
At the same time that we expand multiplayer events, by looking up
the tags of the given ids from the game config, we also note their
addon_id field if present and write this list to the mp_settings,
which will become an attribute of the [multiplayer] tag for the
scenario. When the scenario is hosted, this will tell other
clients what addons they can download to be able to play.