Revert "Some cleanup of map_labels::set_label."
This reverts commit 90501cf87254d3d87f876beb504f8734535d1dea.

This issue was bisected to this commit. This commit claims to just
be a harmless cleanup commit, but it completely breaks the map
labels, causing MP map labels not to be visible, and also breaking
the coloration. I'm not sure what assumptions in this apparently
inoccuous commit are wrong, but it seems that all of this is fixed
and nothing is lost by just reverting it.