campaignd: Add server* this parameter to server::request_handler type
This gets rid of the need for binding this pointers for every registered
request handler in the handler table, and saves us some negligible
theoretical overhead from boost::bind. As a bonus, the
REGISTER_CAMPAIGND_HANDLER macro is more readable.

Now we pass the current pointer during the actual handler calls instead.