make all lobbies, mp ui screens, and chat handlers use the new code
This commit replaces all appearance of `game_config::sounds::...`
where ... is an mp ui specific sound, with an appropriate call to
the mp_ui_sounds code. This makes sure that the preferences are
checked appropriately and notifications are distributed in all cases
appropriate.

This also obsoletes the "lobby sounds" advanced preference, it
doesn't do anything anymore. Its removal is pending.