fix bug 21400: unit::absolute_image being used for unit animations
absolute_image defaults to be the image_icon of a unit, it is
properly used in the sidebar or in relation to portraits of a unit
it only takes the value of "image" when there is no image_icon.

prior to this commit, absolute_image was being used a default
value for animation frames. for many units this is okay, and results
in the base frame being used, but in the cases highlighted by
doofus1 it's broken and results in a portrait being blitted onto
the game map, when image would be a better alternative. This commit
replaces the uses of this in the unit, with either the unit's image
or image_icon, but prioritizing image if both are available.