Merge branch 'boost_filesystem' of https://github.com/AI0867/wesnoth into AI-boost_filesystem_3
Conflicts:
	.travis.yml
	SConstruct
	src/campaign_server/campaign_server.cpp
	src/create_engine.cpp
	src/editor/map/map_context.cpp
	src/game_config_manager.cpp
	src/gui/dialogs/screenshot_notification.cpp
	src/gui/dialogs/wml_error.cpp
	src/savegame.cpp

Conflicts:
	.travis.yml
	SConstruct
	src/save_index.cpp

The second conflict list is from the cherry-pick, the
first is from gfgtdf's original merge.

For me there was additional conflict in
filesystem_boost.cpp, because the
"treat_failure_as_error" flag does not exist in 1.12
branch, it was introduced to accomodate unit tests.

Additionally, dirsize was added even though its not
needed in 1.12 (I removed it), and the change

remove -> filesystem::delete_file

which appear in save_index.cpp in master needed to be
recreated in savegame.cpp in 1.12.