disable auto-backtrace for C++ unit tests, try to get stderr again
We have some problems where the test executable is not generating
all the right output on travis, it seems not to be getting stdout
or stderr or something. The test executable was being run through
gdb at all times, with commands to backtrace on case of crash.

I'm not sure if gdb is eating the output or what, but for now I
am disabling the feature in the hopes to get the error messages
again.