Xcode project: update boost and SDL libraries
This updates to using boost 1.54 (also adding a new library:
libboost_filesystemw.dylib) and SDL_mixer 1.2.12 (plus the other SDL
frameworks, but their upgrade is not essential), in order to work with
the new filesystem code that will be implemented shortly.
This also means that a new Mac compile package is needed.