New Ancient Lich baseframe and updated animations by shadowm
This version of the commit removes the orphaned melee attack animation
frame, which is superseded by a 2-frame set. The 1.12.x version of this
commit keeps the file as a copy of the new melee attack animation frame
for compatibility with existing add-ons.

Animation WML and art reference taken from the existing L3 Lich.