when zooming in, use scale_surface_sharp for units, terrain
Previously we were using scale surface. The result is that sprites
look blurry, terrain looks blurry, and for fractional zoom values
"hex tearing" artifacts are more visible. It seems that zookeeper
and others also prefer a more traditional, just-give-me-the-pixels
look anyways.

lipk even thought that we had been using scale_surface_sharp on
units "since forever", so here we treat the fact that we weren't
as a bug.