change preprocessor symbols controlling SDL MIXER min version
Previously, the code that requires SDL_mixer v 1.2.12 was guarded
by #ifdef PANDORA, since this was the only platform intended to
use this. However, it turns out that OS X < 10.5 has a hard time
getting this library as well.

Instead of branching by platform, here we just check the version
defind by the SDL_mixer header, and include the code appropriate.

We rely on the scons / cmake / project file to fail any build which
should be done with the recent version.