remove "network_ai" and "network" controllers
and instead have a field "is_networked" in [side]. The plan is to remove
"is_networked" and to replace it with a 'controller_client_id ==
local_controller_client_id' check (names like controller_client_id might
change).

This might break existent wml that uses side controller= to check
whether a side is networked, the plan is to add a lua function
wesnoth.is_networked(side). These changes in wml syntax are not side
effects, they are inteded. The reason is, that this makes it easier to
play Sp scenarios that doesn't think about "network_ai" and "network"
controllers in mp, and it disallows stuff like [modify_side]
controller="network" which always causes bugs.