remove "network_ai" and "network" controllers 2
we now also removed the "is_networked" attribute and replaced it with a
controller_client_id == local_controller_client_id check. With this the
contents of [side] should be the same for every side, especially the
server doesn't need to send different [side] tags to different clients.
And we don't have to do serversided controller tweaks anymore.