Upgrade bundled Lua library to version 5.2.3 (bug #22893)
This fixes several bugs in the Lua interpreter:

 * http://www.lua.org/bugs.html#5.2.1
 * http://www.lua.org/bugs.html#5.2.2
 * http://www.lua.org/bugs.html#5.2.3

Released on December 2013, this is the latest stable version as of the
writing of this patch.

This commit consists of a merge of the full diff between versions 5.2.0
and 5.2.3, with paths fixed (.c vs. .cpp). I only had to manually
reconcile the following two commits:

 * f5e673e6443220c57e40cea31d430870b4f73925
   Replace all kinds of the constant pi representations.
 * 299a29f99a84767731dbadd540712d1a37e5e10d
   Fix an off-by-one past-the-end buffer read in lua