fix issue where wml unit tests run in the presence of addon WML
Any scenario of type "test", indeed the generation of the  entire
"TEST" game-config mode, skips all add-ons.

When tests are run, we want them to run in a clean place where
errors and warnings from add-ons don't cause test failures. This
seems to be the simplest way to achieve this right now.

A possible drawback is that if a user defines their own wml unit
tests, they need to move them to their data/test/ folder and
cannot put them in their add-on or they won't be found.

See irc discussion:
http://www.wesnoth.org/irclogs/2014/11/%23wesnoth-dev.2014-11-06.log