Revert "disable "alpha thresholding" in our bilinear interpolation algo"
This reverts commit d73a117278a98921be5915a2eb90675e2cfca0d8.

This is being reverted because there is still yet more bad behavior
in our alpha channel when scaling images, for instance this causes
the terrain overlays to disappear when mousing over terrains with
one of the editor terrain tools.

The reason is that our "scale image" routine is broken and doesn't
weight color contributions by alpha. It's not apporpriate to get
rid of the hacks like alpha thresholding until that has been fixed
and that can't be fixed on 1.12 right now.