"Anonymous local player" -> "Anonymous player"
this string is displayed by non hosts too, so "local" is not true since
it's not local from the other clients point of view.
Also the previous string was too large for a unit name in the sidebar.
Note that this does not effect the "Local player" string in the hosts mp
connect controller options.