use sides_.size() instead of MAX_PLAYES
we usualy want to use that number to index the sides_ vector so
comparing against sides_.size() makes more sense.
We now also use side_index for the index of the side in sides_ and
side_num for the value of side_cfg["side"]
I also put some additional checks.